requirements.txt
django-leaflet
drf-yasg
django-crontab
load_dotenv
django-debug-toolbar
psycopg2-binary
matplotlib
Faker